Added the TaskEdit activity. For now it can only delete tasks
  ---TaskEdit---
    >The activity has the same layout as TaskAdd, albeit a different menu which includes a delete button
    >When an item is pressed* on the List, TaskEdit is launched and all the data of the item are passed to it including its id in the database
    >When the delete button is clicked the id in the database is passed back to TaskList and the item is deleted by the TaskList activity.

  ---OTHER---
    >The task views now also have the due date of the item they represent as a tag with a Date object. The tag has id due_date_tag
       and the tag carrying the database id of the task has id database_task_id
    >Small refactorings

*To make the list entries clickable as wholes, the checkboxes had to be made non-focusable. This doesn't change the behaviour of the
  program regarding those checkboxes.

+++ app/src/main/java/mpardalos/org/homeworkmanager/TaskDatabaseHelper.java
@@ -5,6 +5,7 @@ import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteQueryBuilder;
+import android.util.Log;
 
 import com.readystatesoftware.sqliteasset.SQLiteAssetHelper;
 
@@ -144,8 +145,18 @@ public class TaskDatabaseHelper extends SQLiteAssetHelper {
         db.delete(TASKS_TABLE, null, null);
     }
 
+    /**
+     * Delete a task from the database
+     *
+     * @param taskId the _id of the task to be deleted. If it is -1 then nothing is deleted
+     */
     public void deleteTask(int taskId) {
-        SQLiteDatabase db = getWritableDatabase();
-        db.delete(TASKS_TABLE, "_id=" + String.valueOf(taskId), null);
+        if (!(taskId == -1)) {
+            SQLiteDatabase db = getWritableDatabase();
+            db.delete(TASKS_TABLE, "_id=" + String.valueOf(taskId), null);
+            Log.d("Deleted task. Id", String.valueOf(taskId));
+        } else {
+            Log.d("deleteTask", "Not deleting any task");
+        }
     }
 }
